Refactoring the ConeRayEsimator to be modular
The ConeRayEstimator is made a component and the data collection and
computation are seperated.

With the refactor, it now makes less sense to have the process not be
a component as opposed to it being an independent class which is then
wrapped in another component.

The data collection is deisnged to be their own components as they
could be useful in other situations. The computation is implemented as
reference which can be added using the SerializedReference dropdown.