added a pill interactor
uses the 3 axis formulation for an interactor that is closer to the
shape of the thumb. Produces cones via the estimation process that look
and seem to feel much better and less scattered, but the performance
hasn't been objectively tested. The pill itself could also be generated
on the basis of the measurement of the user's thumb, but currently the
joint follower does not expose enough options to fine tune the
positioning and orientation of the pill

Co-authored-by: Omang Baheti <omangbaheti@gmail.com>