fix incorrect calibration warnings
the warnings produced were happening for phalanges which were not
calibrated for. We already have the editor warning for that. If no data
was collected for a phalange at all, then the "calibration went wrong"
shouldn't be fired. This fixes that by ensuring at least one frame with
one ray interacted with the relevant phalange