Refresh cache generating cone ray angles and on OnEnable of cone
Add a call to RefreshCache on estimatedConeRayAngles after its update
in ConeRayEstimator to ensure cached data gets correctly updated. Add
a public RefreshCache method to HPUIInteractorConeRayAngles and invoke
it in OnEnable to guarantee the cache is refreshed when the interactor
component is enabled or if angles are updated post-creation/loading.